+++ backend/src/main/java/com/empire/store/DatastoreClient.java
@@ -1,6 +1,7 @@
 package com.empire.store;
 
 import com.empire.Nation;
+import com.empire.Orders;
 import com.empire.svc.LoginKey;
 import com.empire.svc.Player;
 import java.util.Set;
