+++ backend/src/main/java/com/empire/store/EntryServlet.java
@@ -2,6 +2,7 @@ package com.empire.store;
 
 import com.empire.Character;
 import com.empire.Nation;
+import com.empire.Orders;
 import com.empire.StartWorldGson;
 import com.empire.svc.LoginCache;
 import com.empire.svc.Player;
